Fix RA Parking/Homing and Slew Limits (#165)
**V1.11.0 - Updates**
- Tracking now stops automatically when the end of the RA ring is reached (using a configurable limit)
- Syncing no longer changes where the firmware expects "home" to be for both DEC and RA
- Fixed RA limits to allow proper meridian flips when slewing
- Fixed parking/return-to-home functionality
- Fixed potential issue of DEC/RA flipping incorrectly when in the southern hemisphere
- Added configuration option for DEC limits
- Fixed a few bugs related to integer overflows and rounding.
- Fixed a bug in RA Autohoming
- Added option for external debugging on separate serial port to allow debugging while using LX200 control
- Updated Polaris coordinates
- Cleaned up logging some more

+++ src/DayTime.cpp
@@ -145,9 +145,9 @@ void DayTime::set(const DayTime &other)
 }
 
 // Add hours, wrapping days (which are not tracked)
-void DayTime::addHours(int deltaHours)
+void DayTime::addHours(float deltaHours)
 {
-    totalSeconds += (long) deltaHours * 3600L;
+    totalSeconds += long(deltaHours * 3600L);
     checkHours();
 }
